2-strncpy.c

+++ 0x06-pointers_arrays_strings/2-strncpy.c
@@ -1,8 +1,15 @@
 #include "main.h"
+/**
+ * *_strncpy - find the length of a string
+ * @dest: pointer to the string
+ * @src: pointer
+ * @n: int
+ * Return: destination
+ */
 
 char *my_strncpy(char *dest, const char *src, size_t n)
 {
-size_t i;
+int i;
 
 for (i = 0; i < n && src[i] != '\0'; i++)
 {
@@ -12,5 +19,5 @@ dest[i] = src[i];
 for (; i < n; i++)
 dest[i] = '\0';
 
- return (dest);
+return (dest);
 }
